fix: fix the WAL hole problem
